Change all prices and alphas from float to double

+++ ggc-core/src/ggc/DerivedProduct.java
@@ -3,9 +3,9 @@ import java.util.ArrayList;
 
 public class DerivedProduct extends Product{
     private ArrayList<RecipeComponent> _recipe;
-    private float _alpha;
+    private double _alpha;
 
-    public DerivedProduct(String name, ArrayList<RecipeComponent> recipe, float alpha){
+    public DerivedProduct(String name, ArrayList<RecipeComponent> recipe, double alpha){
         super(name);
         _recipe = recipe;
         _alpha = alpha;
